% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_ts.R
\name{change_ts}
\alias{change_ts}
\title{Change the input time series from a SaItem}
\usage{
change_ts(ts, sa_item)
}
\arguments{
\item{ts}{the new \code{\link[stats]{ts}} object.}

\item{sa_item}{the SaItem to modify.}
}
\value{
a SaItem
}
\description{
Function change the input time series from a SaItem
}
