% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_ws.R
\name{copy_ws}
\alias{copy_ws}
\title{Copy a WS}
\usage{
copy_ws(ws_name, path_ws, new_path = tempdir())
}
\arguments{
\item{ws_name}{the name of the WS}

\item{path_ws}{the path to the folder containing the WS
(the XML file + the WS folder)}

\item{new_path}{the path to the new folder which will contains the WS
(the XML file + the WS folder)}
}
\value{
TRUE
}
\description{
Copy a WS
}
\examples{

# Déplacement d'un WS dans un environnement temporaire
rjdworkspace:::copy_ws(
  ws_name = "ws_output",
  path_ws = file.path(system.file("extdata", package = "rjdworkspace"), "WS")
)

}
