% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{set_comment}
\alias{set_comment}
\title{Change comment}
\usage{
set_comment(x, comment)
}
\arguments{
\item{x}{the sa_item where to change the comments.}

\item{comment}{the new comment.}
}
\value{
a new sa_item.
}
\description{
Function to change the comments of a sa_item object
}
