% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_series.R
\name{transfer_series}
\alias{transfer_series}
\title{Transfer_series}
\usage{
transfer_series(
  ws_from,
  ws_to,
  selected_series,
  pos_mp_from,
  pos_mp_to,
  name_mp_from,
  name_mp_to,
  print_indications = FALSE,
  create_mp = TRUE,
  replace_series = TRUE
)
}
\arguments{
\item{ws_from}{The workspace containing the additionnal series}

\item{ws_to}{The workspace to add series to}

\item{selected_series}{The vector containing the series-to-update's names.}

\item{pos_mp_from}{The position of the multiprocessing to transfer the series from}

\item{pos_mp_to}{The position of the multiprocessing to transfer the series to}

\item{name_mp_from}{The name of the multiprocessing to transfer the series from (optional)}

\item{name_mp_to}{The name of the multiprocessing to transfer the series to (optional)}

\item{print_indications}{A boolean to print indications on the processing status (optional)}

\item{create_mp}{A boolean to create a new multiprocessing if not existing (optional)}

\item{replace_series}{A boolean to replace existing series (optional)}
}
\value{
the \code{workspace} ws_to augmented with series present in ws_from and not already in ws_to
}
\description{
To copy & paste series from one workspace to another
}
\details{
The transfer will fail if:
- \code{name_mp_from} doesn't exist
- \code{pos_mp_from} < 0 or exceed the maximum number of SAP
- \code{pos_mp_to} < 0 or exceed the maximum number of SAP
- The arguments \code{pos_mp_from} and \code{name_mp_from} are refering to differents objects.
- The arguments \code{pos_mp_to} and \code{name_mp_to} are refering to differents objects.

If \code{name_mp_to} and \code{pos_mp_to} are unspecified, the update will be performed using
the workspaces' first SAProcessing (same for the SAP from).
However if the informations of one on the two SAP (from or to) are specified (name or position), they will be attributed by default to the other worskpace.
}
\examples{

library("RJDemetra")

path_ws_from <- file.path(system.file("extdata", package = "rjdworkspace"),
                          "WS/ws_input.xml")
path_ws_to <- file.path(system.file("extdata", package = "rjdworkspace"),
                        "WS/ws_output.xml")
 
ws_input <- RJDemetra::load_workspace(path_ws_from)
ws_output <- RJDemetra::load_workspace(path_ws_to)

# Existing MP
transfer_series(ws_from = ws_input, 
                ws_to = ws_output, 
                name_mp_from = "SAProcessing-1", 
                name_mp_to = "SAProcessing-1", 
                print_indications = TRUE)
                
transfer_series(ws_from = ws_input, 
                ws_to = ws_output, 
                pos_mp_from = 1, 
                pos_mp_to = 1, 
                print_indications = TRUE)

# Existing series
transfer_series(ws_from = ws_input, ws_to = ws_output, 
                pos_mp_from = 2, 
                pos_mp_to = 2, 
                print_indications = TRUE, 
                replace_series = FALSE)
transfer_series(ws_from = ws_input, ws_to = ws_output, 
                pos_mp_from = 2, 
                pos_mp_to = 2, 
                print_indications = TRUE, 
                replace_series = TRUE)

# Missing MP
# transfer_series(ws_from = ws_input, ws_to = ws_output, 
#                 name_mp_from = "SAProcessing-1", 
#                 name_mp_to = "New-SAProcessing-from-R", 
#                 print_indications = TRUE, 
#                 create = FALSE)

transfer_series(ws_from = ws_input, ws_to = ws_output, 
                name_mp_from = "SAProcessing-1", 
                name_mp_to = "New-SAProcessing-from-R", 
                print_indications = TRUE, 
                create = TRUE)

RJDemetra::save_workspace(workspace = ws_output, file = path_ws_to)

}
