% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_comment}
\alias{get_comment}
\title{Extract comments}
\usage{
get_comment(x)
}
\arguments{
\item{x}{the object where to get the comments.}
}
\description{
Function to extract the comments of a workspace
}
