% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip_sa_item.R
\name{manipulate_sa_item}
\alias{manipulate_sa_item}
\alias{remove_sa_item}
\alias{remove_all_sa_item}
\alias{replace_sa_item}
\alias{add_new_sa_item}
\title{Manipulate SaItems}
\usage{
remove_sa_item(x, pos = 1)

remove_all_sa_item(x)

replace_sa_item(x, pos = 1, sa_item)

add_new_sa_item(x, sa_item)
}
\arguments{
\item{x}{the multiprocessing.}

\item{pos}{the index of the SaItem to remove or to replace.}

\item{sa_item}{SaItem object.}
}
\description{
Functions to remove/replace/add a SAItem from/to a multiprocessing
}
