% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{remove_sa_item}
\alias{remove_sa_item}
\alias{remove_all_sa_item}
\title{Remove SaItems}
\usage{
remove_sa_item(x, pos = 1)

remove_all_sa_item(x)
}
\arguments{
\item{x}{the multiprocessing.}

\item{pos}{the index of the SaItem to remove.}
}
\description{
Functions to remove SAItem from a multiprocessing
}
