% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip_sa_item.R
\name{set_spec}
\alias{set_spec}
\title{Set the specification of a SaItem}
\usage{
set_spec(sa_item, spec)
}
\arguments{
\item{sa_item}{a \code{"sa_item"} object.}

\item{spec}{the object into which the new specification is extracted/stored.}
}
\value{
a new \code{"sa_item"} with the new specification
}
\description{
Function to set the specification of a \code{"sa_item"}.
}
\examples{
sa_x13 <- jx13(ipi_c_eu[, "FR"])
sa_ts <- jtramoseats(ipi_c_eu[, "FR"])

wk <- new_workspace()
mp1 <- new_multiprocessing(wk, "sa1")
add_sa_item(wk, "sa1", sa_x13, "tramo seats")
sa1 <- get_object(mp1, 1)

new_sa <- set_spec(sa1, sa_ts)
# The first sa_item is now seasonally adjusted with TRAMO-SEATS
replace_sa_item(mp1, 1, new_sa)

}
