% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_ts.R
\name{set_ts}
\alias{set_ts}
\title{Change the input time series from a SaItem}
\usage{
set_ts(sa_item, ts)
}
\arguments{
\item{sa_item}{the \code{sa_item} to modify.}

\item{ts}{the new \code{\link[stats]{ts}} object.}
}
\value{
a \code{sa_item}
}
\description{
Function change the input time series from a SaItem
}
\examples{
sa_x13 <- jx13(ipi_c_eu[, "FR"])

wk <- new_workspace()
mp1 <- new_multiprocessing(wk, "sa1")
add_sa_item(wk, "sa1", sa_x13, "X13")
sa1 <- get_object(mp1, 1)

# To create a new sa_item change the input time series
new_sa_item <- set_ts(sa1, ipi_c_eu[, "BE"])
# To replace the first model by the new one
replace_sa_item(mp1, 1, new_sa_item)
}
