% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_series.R
\name{transfer_series}
\alias{transfer_series}
\title{Transfer_series}
\usage{
transfer_series(ws1, ws2, mp_to = NA, mp_from = NA, print_indications = FALSE)
}
\arguments{
\item{ws1}{The workspace to add series to}

\item{ws2}{The workspace containing the additionnal series}

\item{mp_to}{The name of the multiprocessing to transfer the series to}

\item{mp_from}{The name of the multipricessing to transfer the series from}

\item{print_indications}{A boolean to print indications on the processing status (optional)}
}
\value{
the \code{workspace} ws1 augmented with series present in ws2 and not already in ws1
}
\description{
To copy&paste series from one workspace to another
}
\details{
If \code{mp_to} and \code{mp_from} are unspecified, the update will be performed using
the workspaces' first SAProcessing. If only one SAProcessing name is specified but the workspace
it refers to is not (e.g. argument \code{SAP1} instead of \code{mp_from="SAP1"}), it will be
attributed by default to the first worskpace (\code{mp_to="SAP1"}).
}
\examples{
\dontrun{replace_series(ws1, ws2, "SAProcessing-1", "MP2", TRUE)}
}
