% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{update_metadata}
\alias{update_metadata}
\alias{update_metadata_roughly}
\title{Update the metadata from a workspace}
\usage{
update_metadata(workspace1, workspace2)

update_metadata_roughly(workspace1, workspace2)
}
\arguments{
\item{workspace1}{Workspace that contains the new metadata.}

\item{workspace2}{Workspace to update.}
}
\description{
Functions to update the metadata of a workspace by those contained in another one
}
\details{
\code{update_metadata()} checks the multiprocessings and SaItems' names within
the two workspaces before updating workspace2's metadata. \code{update_metadata_roughly()} does not do any
checks: \code{workspace2}'s first multiprocessing's first SaItem metadata
is updated with \code{workspace1}'s first multiprocessing's first SaItem metadata.
Both functions create and return a new workspace containing the updated series.
}
\examples{
\donttest{
workspace1 <- load_workspace("D:/test_metadata/reference.xml")
compute(workspace1)
workspace2 <- load_workspace("D:/test_metadata/ws_wk.xml")
compute(workspace2)

updated_workspace <- update_metadata_roughly(workspace1, workspace2)
save_workspace(updated_workspace, "D:/test_metadata/ws_wk_metadata.xml")

updated_workspace <- update_metadata(workspace1, workspace2)
save_workspace(updated_workspace, "D:/test_metadata/ws_wk_metadata.xml")
}

}
