% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_path.R
\name{update_path}
\alias{update_path}
\title{Update the path to the raw series file}
\usage{
update_path(ws_xml_path, raw_data_path, pos_mp, pos_sa_item)
}
\description{
Fonction de changement de chemin (pour les données brutes)
Cette fonction fonctionne pour les formats csv, xls et xlsx

Function to change the path to the raw series file in a workspace
}
