% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{update_path}
\alias{update_path}
\title{Update the path to the raw series file}
\usage{
update_path(ws_path, raw_data_path, param = 8, print_log = FALSE)
}
\arguments{
\item{ws_path}{the path to the workspace}

\item{raw_data_path}{the path to the raw data file}

\item{param}{the number of lines between a series name and the path to its raw file in the xml file. The default value is 8.}

\item{print_log}{a boolean to indicate whether to print progress indications.}
}
\value{
the eventual list of series from the workspace that weren't found in the raw series file
}
\description{
Function to change the path to the raw series file in a workspace
}
\details{
\itemize{
\item The paths' translation into ASCII is done by the function.
\item The function operates on a workspace's first SAProcessing (ie. on a whole workspace when it contains one SAP).
\item The raw series file must be a csv file.
\item The paths must be "full" (as opposed to relative to the setwd directory, for example).
}
}
\examples{

path_to_ws <- file.path(system.file("extdata", package = "rjdworkspace"),
                        "WS/ws_example_path.xml")
                        
path_to_raw_data <- file.path(system.file("extdata", package = "rjdworkspace"),
                              "data_file.csv")
                              
# Minimal syntax
update_path(ws_path = path_to_ws, raw_data_path = path_to_raw_data)

}
